requirements.txt
# ================================
# CORE ENV
# ================================
python-dotenv==1.0.1

# ================================
# PDF PROCESSING
# ================================
pymupdf==1.24.9
pdf2image==1.17.0
pillow==10.4.0

# ================================
# OCR (EASYOCR STACK)
# ================================
easyocr==1.7.2
opencv-python-headless==4.6.0.66
numpy==1.26.4
scipy==1.11.4
scikit-image==0.22.0
python-bidi==0.4.2
pyclipper==1.3.0.post5
Shapely==2.0.4
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# ================================
# DATA HANDLING
# ================================
pandas==2.2.2

# ================================
# DATABASE
# ================================
pymongo==4.8.0

# ================================
# EMBEDDINGS
# ================================
sentence-transformers==2.7.0
transformers==4.44.2

# ================================
# LANGCHAIN ECOSYSTEM
# ================================
langchain==0.2.14
langchain-core==0.2.35
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langchain-huggingface==0.0.3
langchain-groq==0.1.9

# ================================
# STREAMLIT UI
# ================================
streamlit==1.37.1

# ================================
# UTILITIES
# ================================
tqdm==4.66.4
requests==2.32.3
PyYAML==6.0.2
